fixed: when non ASCII character in .profile on CJK Windows
I use GWSL.
And, I got error.

~~~console
2021-12-07 00:57:32,949 - GWSL 1.4.0 - ERROR - Exception occurred - Error in Mainloop
Traceback (most recent call last):
  File "manager.py", line 4451, in <module>
  File "manager.py", line 4061, in draw
  File "manager.py", line 4014, in apper
  File "manager.py", line 2266, in app_launcher
  File "manager.py", line 2805, in spawn_n_run
  File "wsl_tools.py", line 191, in profile
UnicodeDecodeError: 'cp932' codec can't decode byte 0x86 in position 586: illegal multibyte sequence
~~~

I research.

I found the cause of this error to be the following.

In Japanese Windows, the default system character set is cp932 (Shift_JIS).
Therefore, Python also basically tries to handle strings as cp932.
However, in GNU/Linux, the character encoding of files is basically ASCII or UTF-8.
At least, it will be more than cp932.
Therefore, when reading `.profile', it is better to expect UTF-8 and read it with a higher success rate.
Therefore, I basically rewrote the wsl.exe process to assume that the return value of the process is UTF-8 and decode it.