Refactor looping logic for clarity
The looping logic for the profiles requires that the GUI continues to
update until it is either closed or options are submitted. I moved both
of these requirements for the loop out of the loop body and into the
loop condition, thereby reducing the complexity of the logic flow. It
was also necessary to make sure that there were options, before
returning them, so that the behavior of the public API would not change.